requirements.txt
anyio==3.5.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bitrix24-rest==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.12
dadata==21.10.1
Django==4.0.4
django-bootstrap-v5==1.0.11
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
oyaml==1.0
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
